refactor(server): extract worker command service
