+++ scripts/max_lines_allowlist.txt
@@ -41,7 +41,7 @@ server/proliferate/server/cloud/runtime/bootstrap.py 922 server structure migrat
 server/proliferate/server/cloud/runtime/provision.py 1688 server structure migration debt for oversized domain-adjacent file
 server/proliferate/server/cloud/runtime_config/domain/resolver.py 611 server structure migration debt for oversized pure-domain file
 server/proliferate/server/cloud/runtime_config/service.py 1095 server structure migration debt for oversized service file
-server/proliferate/server/cloud/worker/service.py 1290 server structure migration debt for oversized service file
+server/proliferate/server/cloud/worker/service.py 912 server structure migration debt for oversized service file
 server/proliferate/server/cloud/workspaces/models.py 934 server structure migration debt for oversized API models file
 server/proliferate/server/cloud/workspaces/service.py 2643 server structure migration debt for oversized service file
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
